% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xyzt.R
\name{create_xyzt}
\alias{create_xyzt}
\title{Create xyzt particle set}
\usage{
create_xyzt(x, y, z, t, prj = NULL)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{z}{z}

\item{t}{t}

\item{prj}{prj}
}
\value{
xyzt particle set
}
\description{
Create xyzt particle set
}

