% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Track a particle set}
\usage{
track(xyzt, tstep, nsteps, dis, bas, hed, bud = NULL,
  darcy = convert_bud_to_darcy(bud, dis = dis), dsp = NULL, terminals,
  mpmain = NULL, parallel = FALSE, cores = parallel::detectCores(),
  type = "labolle")
}
\description{
Track a particle set
}

